# src/backend/openapi_server/models/update_turn_request.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class UpdateTurnRequest(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, uid=None):  # noqa: E501
        """UpdateTurnRequest - a model defined in OpenAPI

        :param uid: The uid of this UpdateTurnRequest.  # noqa: E501
        :type uid: str
        """
        self.openapi_types = {
            'uid': str
        }

        self.attribute_map = {
            'uid': 'uid'
        }

        self._uid = uid

    @classmethod
    def from_dict(cls, dikt) -> 'UpdateTurnRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UpdateTurnRequest of this UpdateTurnRequest.  # noqa: E501
        :rtype: UpdateTurnRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def uid(self) -> str:
        """Gets the uid of this UpdateTurnRequest.


        :return: The uid of this UpdateTurnRequest.
        :rtype: str
        """
        return self._uid

    @uid.setter
    def uid(self, uid: str):
        """Sets the uid of this UpdateTurnRequest.


        :param uid: The uid of this UpdateTurnRequest.
        :type uid: str
        """
        if uid is None:
            raise ValueError("Invalid value for `uid`, must not be `None`")  # noqa: E501

        self._uid = uid