# src/backend/openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: API für das JassSpiel Backend
  title: JassSpiel API
  version: v1.0.0
servers:
- description: Lokaler Entwicklungsserver
  url: http://localhost:3000
...
paths:
...
  /rounds/{rid}/first-card:
    get:
      operationId: rounds_rid_first_card_get
      parameters:
      - explode: false
        in: path
        name: rid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jasskarte"
          description: Jasskarten-Objekt oder null.
      summary: Holt das Objekt der ersten gespielten Karte in einer Runde.
      tags:
      - Rounds
      x-openapi-router-controller: openapi_server.controllers.rounds_controller
  /rounds/{rid}/first-card-cid:
    get:
      operationId: rounds_rid_first_card_cid_get
      parameters:
      - explode: false
        in: path
        name: rid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rounds_rid_first_card_cid_get_200_response"
          description: CID der ersten Karte oder leer.
      summary: Holt die CID der ersten gespielten Karte in einer Runde.
      tags:
      - Rounds
      x-openapi-router-controller: openapi_server.controllers.rounds_controller
  /rounds/{rid}/played-cards:
    get:
      operationId: rounds_rid_played_cards_get
      parameters:
      - explode: false
        in: path
        name: rid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Jasskarte"
                type: array
          description: Liste der gespielten Karten.
      summary: Holt alle gespielten Karten einer Runde.
      tags:
      - Rounds
      x-openapi-router-controller: openapi_server.controllers.rounds_controller
  /rounds/{rid}/plays:
    post:
      operationId: rounds_rid_plays_post
      parameters:
      - explode: false
        in: path
        name: rid
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPlayRequest"
        required: true
      responses:
        "201":
          description: Spielzug erfolgreich hinzugefügt.
      summary: Fügt einen Spielzug (gespielte Karte) zu einer Runde hinzu.
      tags:
      - Rounds
      x-openapi-router-controller: openapi_server.controllers.rounds_controller
  /rounds/{rid}/turn:
    get:
      operationId: rounds_rid_turn_get
      parameters:
      - explode: false
        in: path
        name: rid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/games_gid_next_player_uid_get_200_response"
          description: UID des Spielers am Zug.
      summary: "Holt die UID des Spielers, der aktuell am Zug ist."
      tags:
      - Rounds
      x-openapi-router-controller: openapi_server.controllers.rounds_controller
    put:
      operationId: rounds_rid_turn_put
      parameters:
      - explode: false
        in: path
        name: rid
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTurnRequest"
        required: true
      responses:
        "200":
          description: Spieler am Zug erfolgreich aktualisiert.
      summary: "Aktualisiert, welcher Spieler am Zug ist."
      tags:
      - Rounds
      x-openapi-router-controller: openapi_server.controllers.rounds_controller
  /rounds/{rid}/winner:
    put:
      operationId: rounds_rid_winner_put
      parameters:
      - explode: false
        in: path
        name: rid
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWinnerRequest"
        required: true
      responses:
        "200":
          description: Gewinner erfolgreich aktualisiert.
      summary: Aktualisiert den Gewinner einer Runde.
      tags:
      - Rounds
      x-openapi-router-controller: openapi_server.controllers.rounds_controller
...
components:
  schemas:
    Jasskarte:
      example:
        symbol: symbol
        path: path
        cardtype: cardtype
        cid: cid
      properties:
        cid:
          description: Eindeutige ID der Karte.
          title: cid
          type: string
        symbol:
          description: "Farbe/Symbol der Karte (z.B. Schella, Rose)."
          title: symbol
          type: string
        cardtype:
          description: "Typ der Karte (z.B. Ass, König, 6)."
          title: cardtype
          type: string
        path:
          description: Pfad zum Bild der Karte.
          title: path
          type: string
      title: Jasskarte
      type: object
...
    AddPlayRequest:
      example:
        uid: uid
        cid: cid
      properties:
        uid:
          description: "UID des Spielers, der die Karte spielt."
          title: uid
          type: string
        cid:
          description: CID der gespielten Karte.
          title: cid
          type: string
      required:
      - cid
      - uid
      title: AddPlayRequest
      type: object
...
    UpdateWinnerRequest:
      example:
        winnerUid: winnerUid
      properties:
        winnerUid:
          title: winnerUid
          type: string
      required:
      - winnerUid
      title: UpdateWinnerRequest
      type: object
    UpdateTurnRequest:
      example:
        uid: uid
      properties:
        uid:
          title: uid
          type: string
      required:
      - uid
      title: UpdateTurnRequest
      type: object
...
    games_gid_next_player_uid_get_200_response:
      example:
        uid: uid
      properties:
        uid:
          title: uid
          type: string
      title: games_gid_next_player_uid_get_200_response
      type: object
...
    rounds_rid_first_card_cid_get_200_response:
      example:
        cid: cid
      properties:
        cid:
          nullable: true
          title: cid
          type: string
      title: rounds_rid_first_card_cid_get_200_response
      type: object
...